MonadSTM: split into submodules
Follow a similar module structure as the "stm" package.
