+++ io-classes/src/Control/Concurrent/Class/MonadSTM/TArray.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE ExplicitNamespaces #-}
+
+-- | This module corresponds to `Control.Concurrent.STM.TArray` in "stm" package
+--
+module Control.Concurrent.Class.MonadSTM.TArray (type TArray) where
+
+import           Control.Monad.Class.MonadSTM.Internal
