+++ io-classes/src/Control/Concurrent/Class/MonadSTM/TChan.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE ExplicitNamespaces #-}
+
+-- | This module corresponds to `Control.Concurrent.STM.TChan` in "stm" package
+--
+module Control.Concurrent.Class.MonadSTM.TChan
+  ( -- * MonadSTM
+    -- ** TChans
+    type TChan
+    -- * Construction
+  , newTChan
+  , newBroadcastTChan
+  , newTChanIO
+  , newBroadcastTChanIO
+  , dupTChan
+  , cloneTChan
+    -- ** Reading and writing
+  , readTChan
+  , tryReadTChan
+  , peekTChan
+  , tryPeekTChan
+  , writeTChan
+  , unGetTChan
+  , isEmptyTChan
+  ) where
+
+import           Control.Monad.Class.MonadSTM.Internal
