+++ io-classes/src/Control/Concurrent/Class/MonadSTM/TMVar.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE ExplicitNamespaces #-}
+
+-- | This module corresponds to `Control.Concurrent.STM.TMVar` in "stm" package
+--
+module Control.Concurrent.Class.MonadSTM.TMVar
+  ( -- * MonadSTM
+    type TMVar
+  , newTMVar
+  , newEmptyTMVar
+  , takeTMVar
+  , tryTakeTMVar
+  , putTMVar
+  , tryPutTMVar
+  , readTMVar
+  , tryReadTMVar
+  , swapTMVar
+  , isEmptyTMVar
+    -- * MonadLabelledSTM
+  , labelTMVar
+  , labelTMVarIO
+    -- * MonadTraceSTM
+  , traceTMVar
+  , traceTMVarIO
+  ) where
+
+import           Control.Monad.Class.MonadSTM.Internal
